refactor: 파일명 변경 CompanionDogSection2 -> CompanionDogSection
