+++ frontend/src/components/journals/CompanionDogSection.tsx
